Move PatchError to task_patch.rs.
