+++ src/github_issues.rs
@@ -4,10 +4,10 @@ use crate::filter;
 use crate::github::client::Client;
 use crate::github::structs;
 use crate::project::Project as ProjectTrait;
-use crate::provider::{PatchError, Provider as ProviderTrait};
+use crate::provider::Provider as ProviderTrait;
 use crate::task::due_group;
 use crate::task::{DateTimeUtc, State, Task as TaskTrait};
-use crate::task_patch::TaskPatch;
+use crate::task_patch::{PatchError, TaskPatch};
 use chrono::{DateTime, NaiveDate, NaiveDateTime, Utc};
 use ratatui::style::Color;
 use std::any::Any;
