+++ src/gitlab_todo.rs
@@ -4,9 +4,9 @@ use crate::filter;
 use crate::gitlab::client::Client;
 use crate::gitlab::structs;
 use crate::project::Project as ProjectTrait;
-use crate::provider::{PatchError, Provider as ProviderTrait};
+use crate::provider::Provider as ProviderTrait;
 use crate::task::{DateTimeUtc, State, Task as TaskTrait};
-use crate::task_patch::TaskPatch;
+use crate::task_patch::{PatchError, TaskPatch};
 use chrono::{DateTime, NaiveDate, NaiveDateTime, NaiveTime, Utc};
 use ratatui::style::Color;
 use std::any::Any;
