+++ src/task_patch.rs
@@ -65,6 +65,23 @@ impl Clone for TaskPatch {
     }
 }
 
+pub struct PatchError {
+    pub task: Box<dyn TaskTrait>,
+    pub error: String,
+}
+
+impl PatchError {
+    pub fn is_task(&self, task: &dyn TaskTrait) -> bool {
+        self.task.id() == task.id() && self.task.provider() == task.provider()
+    }
+}
+
+impl std::fmt::Display for PatchError {
+    fn fmt(&self, f: &mut std::fmt::Formatter) -> std::fmt::Result {
+        write!(f, "Error patching task with id {}: {}", self.task.id(), self.error)
+    }
+}
+
 #[cfg(test)]
 mod tests {
     use super::*;
